+++ tests/testthat/test-conversion.R
@@ -1,5 +1,11 @@
 context("Test conversion")
 
+setup(RNGversion("3.5.3"))
+teardown({
+  cur_R_version <- trimws(substr(R.version.string, 10, 16))
+  RNGversion(cur_R_version)
+})
+
 test_that("Test against reference results", {
     skip_on_cran()
 
@@ -9,7 +15,7 @@ test_that("Test against reference results", {
 
 
     ## example with a function for SI
-    si <- distcrete("gamma", interval = 1L,
+    si <- distcrete::distcrete("gamma", interval = 1L,
                     shape = 1.5,
                     scale = 2, w = 0)
 
@@ -18,12 +24,12 @@ test_that("Test against reference results", {
 
     ## basic export
     df_1 <- as.data.frame(pred_1)
-    expect_equal_to_reference(df_1, file = "rds/df_1.rds")
+    expect_equal_to_reference(df_1, file = "rds/df_1.rds", update = FALSE)
     expect_equal(as.vector(unlist(df_1[, -1])), as.vector(pred_1))
 
     ## long format
     df_2 <- as.data.frame(pred_1, long = TRUE)
-    expect_equal_to_reference(df_2, file = "rds/df_2.rds")
+    expect_equal_to_reference(df_2, file = "rds/df_2.rds", update = FALSE)
 
 
 })