+++ tests/testthat/test-subset.R
@@ -1,5 +1,11 @@
 context("Test subset")
 
+setup(RNGversion("3.5.3"))
+teardown({
+  cur_R_version <- trimws(substr(R.version.string, 10, 16))
+  RNGversion(cur_R_version)
+})
+
 test_that("Test against reference results - numeric dates", {
     skip_on_cran()
 
@@ -9,7 +15,7 @@ test_that("Test against reference results - numeric dates", {
 
 
     ## example with a function for SI
-    si <- distcrete("gamma", interval = 1L,
+    si <- distcrete::distcrete("gamma", interval = 1L,
                     shape = 1.5,
                     scale = 2, w = 0)
 
@@ -21,9 +27,9 @@ test_that("Test against reference results - numeric dates", {
     subset_3 <- subset(pred_1, to = 15, sim = c(TRUE, FALSE))
 
     expect_identical(pred_1[], pred_1)
-    expect_equal_to_reference(subset_1, file = "rds/subset_1.rds")
-    expect_equal_to_reference(subset_2, file = "rds/subset_2.rds")
-    expect_equal_to_reference(subset_3, file = "rds/subset_3.rds")
+    expect_equal_to_reference(subset_1, file = "rds/subset_1.rds", update = FALSE)
+    expect_equal_to_reference(subset_2, file = "rds/subset_2.rds", update = FALSE)
+    expect_equal_to_reference(subset_3, file = "rds/subset_3.rds", update = FALSE)
     expect_error(subset(pred_1, from = 1, to = 0), "No data retained.")
 
 })
@@ -42,7 +48,7 @@ test_that("Test against reference results - Date dates", {
 
 
     ## example with a function for SI
-    si <- distcrete("gamma", interval = 1L,
+    si <- distcrete::distcrete("gamma", interval = 1L,
                     shape = 1.5,
                     scale = 2, w = 0)
 
@@ -54,9 +60,10 @@ test_that("Test against reference results - Date dates", {
     subset_3 <- subset(pred_1, to = day + 15, sim = c(TRUE, FALSE))
 
     expect_identical(pred_1[], pred_1)
-    expect_equal_to_reference(subset_1, file = "rds/subset_Date_1.rds")
-    expect_equal_to_reference(subset_2, file = "rds/subset_Date_2.rds")
-    expect_equal_to_reference(subset_3, file = "rds/subset_Date_3.rds")
+    expect_equal_to_reference(subset_1, file = "rds/subset_Date_1.rds", update = FALSE)
+    expect_equal_to_reference(subset_2, file = "rds/subset_Date_2.rds", update = FALSE)
+    expect_equal_to_reference(subset_3, file = "rds/subset_Date_3.rds", update = FALSE)
 
 })
+